+++ src/interpreter.ts
@@ -60,6 +60,9 @@ export class Interpreter {
     state.z = z ?? state.z;
 
     currentPath.addPoint(state.x, state.y, state.z);
+    if (pathType === PathType.Extrusion) {
+      job.boundingBox.update(state.x, state.y, state.z);
+    }
   }
 
   g1 = this.g0;
@@ -160,6 +163,9 @@ export class Interpreter {
       py = centerY + arcRadius * Math.sin(currentAngle);
       pz += zStep;
       currentPath.addPoint(px, py, pz);
+      if (pathType === PathType.Extrusion) {
+        job.boundingBox.update(px, py, pz);
+      }
     }
 
     state.x = x || state.x;
@@ -167,6 +173,9 @@ export class Interpreter {
     state.z = z || state.z;
 
     currentPath.addPoint(state.x, state.y, state.z);
+    if (pathType === PathType.Extrusion) {
+      job.boundingBox.update(state.x, state.y, state.z);
+    }
   }
 
   g3 = this.g2;
